+++ src/Adapter/ShutdownHandler/shutdownTask.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {() => Promise<void>} shutdownTask
- */